fix: Visualización de peticiones por parte del Ordenador

+++ models/cumplido_rp.go
@@ -1,19 +1,19 @@
 package models
 
 type CumplidoRp struct {
-	Id 						int
-	ContratoId 				*Contrato
-	PreliquidacionId		*Preliquidacion
-	Cumplido				bool
-	Preliquidado			bool
-	ResponsableIva			bool
-	Dependientes			bool
-	Pensionado				bool
-	InteresesVivienda		int
-	MedicinaPrepagadaUvt	int
-	PensionVoluntaria		int
-	Afc 					int
-	Activo					bool
-	FechaCreacion			string
-	FechaModificacion 		string
+	Id                   int
+	ContratoId           *Contrato
+	PreliquidacionId     *Preliquidacion
+	Cumplido             bool
+	Preliquidado         bool
+	ResponsableIva       bool
+	Dependientes         bool
+	Pensionado           bool
+	InteresesVivienda    float64
+	MedicinaPrepagadaUvt float64
+	PensionVoluntaria    float64
+	Afc                  float64
+	Activo               bool
+	FechaCreacion        string
+	FechaModificacion    string
 }
